index.js

+++ my-service/index.js
@@ -206,7 +206,7 @@ app.delete("/api/reviews/:id", auth, (req, res) => {
         const { id } = req.params
         const review = db.prepare("SELECT * FROM review WHERE id = ?").get(id)
         if (!review) return res.status(404).json({ error: "Отзыв не найден" })
-
+        
         db.prepare('DELETE FROM review WHERE id = ?').run(id)
         return res.status(200).json({ message: 'Deleted successfully' })
     } catch (error) {
@@ -220,7 +220,11 @@ app.delete("/api/books/:id", auth, (req, res) => {
         const { id } = req.params
         const book = db.prepare("SELECT * FROM books WHERE id = ?").get(id)
         if (!book) return res.status(404).json({ error: "Книга не найдена" })
-
+        if (!['admin'].includes(req.user.role) || req.user.id !== Reviews.userId) {
+            return res
+            .status(403)
+            .json({message: 'Доступ запрещен: недостаточно прав'})
+        }
         db.prepare('DELETE FROM books WHERE id = ?').run(id)
         return res.status(200).json({ message: 'Deleted successfully' })
     } catch (error) {
@@ -271,7 +275,11 @@ app.delete("/api/admin/users/:id", auth, (req, res) => {
         const { id } = req.params
         const user = db.prepare("SELECT * FROM users WHERE id = ?").get(id)
         if (!user) return res.status(404).json({ error: "Юзер не найден" })
-
+        if (!['admin'].includes(req.user.role) || req.user.id !== Reviews.userId) {
+            return res
+            .status(403)
+            .json({message: 'Доступ запрещен: недостаточно прав'})
+        }
         db.prepare('DELETE FROM users WHERE id = ?').run(id)
         return res.status(200).json({ message: 'Deleted successfully' })
     } catch (error) {
